python/requirements.txt
rasa-pro
rasa-sdk
discord.py
python-dotenv
